Update proto definitions
I think I found the corresponding file in the anki repo so I was able to
update the definitions to be more accurate.

+++ backend/src/api/anki/data.rs
@@ -1,5 +1,3 @@
-use std::fmt;
-
 use chrono::{DateTime, Utc};
 
 use super::proto_definitions::DeckInfo;
@@ -16,8 +14,8 @@ impl From<DeckInfo> for AnkiData {
         deck.review_card_count();
 
         Self {
-            active_review_count: deck.review_card_count() as u32,
-            new_card_count: deck.new_card_count() as u32,
+            active_review_count: deck.review_card_count(),
+            new_card_count: deck.new_card_count(),
             data_updated_at: Utc::now(),
         }
     }
